requirements.txt
langchain
streamlit
python-dotenv
PyPDF2


-e .